testing: deploy Hashicorp Vault with OCS
With these modifications, it is possible to deploy Hashicorp Vault in
the openshift-storage namespace like other OCS components.

The provided configuration gives a "vault-test" KMS connection in the
"csi-kms-connection-details" ConfigMap. With the matching permissions on
the Hashicorp Vault side, the ServiceAccount used by the Ceph-CSI RBD
provisioner and node-plugin can create/fetch/delete passphrases in the
KMS for encrypting PVs.

Signed-off-by: Niels de Vos <ndevos@redhat.com>