util: convert standardVault object to map
getVaultConfiguration() parses the standardVault object from the
ConfigMap. It then sets the value to the given KMS-ID. Unfortunately
the value is in the form of a standardVault object, where a
map[string]interface{} is expected.

It seems that dataToStandardVault() converts the standardVault object to
the correct format, but does not return it. By addressing this problem,
the configuration from the ConfigMap gets parsed and converted correctly
again.

Signed-off-by: Niels de Vos <ndevos@redhat.com>